fixed all tests
